[Fabric] Refactor events, remove unused code (#422)
* fix: move ReactSliderShadowNode to oldarch

* fix: refactor to RCTModernEventEmitter closes #412

* fix: minimize diff & indentations
