+++ package/android/src/main/java/com/reactnativecommunity/slider/ReactSlidingCompleteEvent.java
@@ -10,7 +10,7 @@ package com.reactnativecommunity.slider;
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.uimanager.events.Event;
-import com.facebook.react.uimanager.events.RCTEventEmitter;
+import com.facebook.react.uimanager.events.RCTModernEventEmitter;
 
 /**
  * Event emitted when the user finishes dragging the slider.
@@ -46,7 +46,7 @@ public class ReactSlidingCompleteEvent extends Event<ReactSlidingCompleteEvent> 
     }
 
     @Override
-    public void dispatch(RCTEventEmitter rctEventEmitter) {
+    public void dispatchModern(RCTModernEventEmitter rctEventEmitter) {
         rctEventEmitter.receiveEvent(getViewTag(), getEventName(), serializeEventData());
     }
